Use a single div for Paper
In a conversation with Hai Nguyen (@hai-cea), he confirmed that the
only reason for creating an inner div was preserving the two shadows.
Since it's possible in CSS to apply two or more shadows to a single
element, such nested structure is in my opinion redundant.
The order of shadows matters: shadows for an element are applied
top to bottom (first one goes on top), we had to put the child's
shadows first.