Removes `mixins/controllable` and replaces functionality with local methods
The `mixins/controllable.js` mixin adds a `getValueLink()` method to components and is currently used in two places:

- Menu
- Tabs

This commit moves the function locally into each of the components so that they don't depend on the `controllable.js` mixin. The functions should be removed when we're ready to officially deprecate valueLinks.