Make classnames a direct dependency
I don't use classnames and it's not the best way to require installing
it from user. Peer depenencies should be used carefully. react,
material-ui are good packages for this purpose. Keeping classnames
in peer depenencies doesn't solve any problem. It's not so big package
to be a dependency bloat. And if one day there will be a different
version then I'm sure it will be a breaking change and users app will be
broken.