[Slider] Improve performance of slider (#13325)
* [Slider] Improve performance of slider

Changing left/right/top/bottom or width/height causes layout which can
be quite expensive. This is not a huge problem on the demo site as it is
relatively simple, but even on it you can see a big difference on the two
synced sliders. As layout cost raises with the complexity of the complete
page this can cause performance problems even with single sliders as well.

This PR replaces layout causing properties with transform and add
will-change so that slider updates can be done without any layout or paint.

Disabled sliders do still cause layout when changed, due to the increased
complexity of creating the "gap" between the track and the thumb. However,
as these are not dragged by the user it is much less of a problem as slow
updates are far less noticeable.

* simplify slider implementation

* remove unit tests that test implementation and are already covered with visual testing

* fix lint error