[docs] Batch changes (#16621)
* [docs] Try out an affiliation link under my name for 1 month to see the potential

I will create an Amazon associate account for Material-UI inc. if this experiment is successful.
For now, it was simpler to use my existing verfied account.
I have no idea about the potential. $10/month? $100/month? $1,000/month?

* [docs] The CSP guide was moved to the advanced styles section

* [docs] Not the most popular in China

* [test] innerRef shouldn't be used anymore

* [docs] Replace property -> prop on the JavaScript sources

I haven't touched the markdowns to reduce the overhead of this effort and to to invalidate the translation.
We gonna have to do it at some point.

* [docs] Sync the typography markdown with the demo

* [test] Use the render = createClientRender convention

* fix build